src/pyproject.toml
[project]
name = "src"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain-text-splitters>=0.3.8",
    "python-dotenv>=1.1.1",
]
